refactor!: Deprecate `includes` in `toObject()`
The custom `includes` argument for the `toObject()` method returned
by the `RecordHandler` for a `Record` is now deprecated. Serializers
cover this need and provide a far more robust API for this use-case.

BREAKING CHANGE: The `toObject()` function returned from the
`RecordHandler` on a `Record` object doesn't accept any arguments
anymore.