feat!: Deprecate model key configuration
Deprecation of the `key` parameter for `Model` definitions. All models
now use the convention of `'id'` being their key name and it being a
string type.
This commit introduces the deprecation of the functionality and API,
along with documentation changes to match. Remaining code related
to key customization will be removed in subsequent commits.
Documentation has been updated accordingly.

Refs: #31
BREAKING CHANGE: Model `key` definitions are deprecated. All model keys
are now conventionally named `'id'` and are of a string type.