feat: Simplify key usage in relationships
As all keys have the same type and name, their usage in relationships
can be vastly simplified to utilize their known properties.
An internal `keyArray` type was introduced for multi-record
relationship fields and all usage of customizable relationship fields
was replaced with the conventional field name (`id`) and string type.

Refs: #31