feat!: Make `fields` definition an object
Update the `fields` argument of the `Model` constructor to expect an
object, keyed by field name, instead of an array of objects. This makes
the field definition more similar to the property definition.
Documentation and REPL have been updated accordingly.

Refs: #25
BREAKING CHANGE: The `fields` key passed to the `Model` constructor now
expects an object indexed on the field name instead of an object array.