refactor: Make all properties act as lazy
Simplify the schema creation logic to benefit from the fact
that the schema is now immutable after creation. This allows
for the internal separation between regular and lazy properties
to be eliminated, using instead a protected getter to retrieve the
schema object on the fly. This change, in turn, allows for
`Model.prototype[$addProperty]` to be privatized and the symbol
to be removed. The `RecordHandler` is also updated to always
retrieve `Schema[$schemaObject]` on the fly when a property is
calculated. Property caching is not affected, as potential cache
invalidation issues predate this change. Trivial performance gains
might be noticed during schema creation.

Refs: #26