refactor: Deprecate `lazyMethods` for serializers
Remove experimental behavior related to `lazyMethods`
definitions in `Serializer` creation. Lazy methods and properties
should be defined on the `Model` level and their values passed
down to the `Serializer` as needed.
This may break undocumented behavior, but is not considered a
breaking change, due to it never being knowingly suggested for use.