refactor!: Call `methods` with `Schema` singleton
As the schema definition is now immutable, all methods defined
for any model can be called safely with the current schema. This
allows for definitions to use the last argument, optionally, to get
the current schema value via `Schema[$schemaObject]`. This,
in turn, allows for `Model[$addMethod]` to be privatized and the
symbol to be removed. Documentation has been updated accordingly.

Refs: #27
BREAKING CHANGE: You can no longer define `lazyMethods` as part
of model definitions. Regular `methods` can now accept a final argument
that represents the current schema.