refactor: Privatize method creation for `Serializer`
As the entirety of the schema is now immutable, the option
to create new methods for `Serializer` instances should not
be available. This only alters internal functionality, making
`Serializer.prototype.addMethod()` private.
This may break undocumented behavior, but is not considered a
breaking change, due to it not being suggested for use.