feat: Optimize performance of `RecordSet`
`RecordSet` querying operations will now create new instances of the
`RecordSet` class with an empty `iterable` and utilize the protected
`RecordSet.prototype[$set]` method to populate the instance. This
minimizes time spent on iterating over the same entries and increases
performance. Both previous and current complexity are `O(n)`, however
the constant `c` has been reduced from `2` down to `1` in many cases.
This should be especially noticeable for scoped queries without a
comparation function, which is one of the most common operations.